Check out latest gitian-builder branch during fetch-inputs
When using git checkout <name>, and <name> does not correspond to a
locally existing branch but does correspond to a remotely existing
branch, a local branch of that name is created, set to the contents of
the remote branch, and then checked out. If a local branch with that
name already exists (which happens when you run git checkout <name> for
a second time) then only the local branch is checked out, without any
regard for the remote content changes. Make sure we actually update to
the remote content.