Download argparse.
It's on a mirror because it 404s half the time. Restarting the download
right away always works. But it often fails the first time you run
fetch-inputs.sh after a while. The googlecode download also has the
property that you have to download the entire file even if your local
copy is newer (that's why the TOOLCHAIN4 download from googlecode is
also disabled), but that's not as big a problem as the failed downloads
because the argparse package is small.