Add gcc, dllwrap, and swig wrappers for Wine.
Python distutils, with the "mingw32" compiler profile, is inflexible. It
hardcodes the names "gcc.exe" and "dllwrap.exe", and passes them fixed
options. It's not possible to control it with environment variables like
CC and CFLAGS.

Before Python 2.7.6, distutils always passes the option -mno-cygwin to
gcc, which causes an error. Additionally, when distutils is run under
Wine, it passes Windows-style paths with backslashes to our unsuspecting
i686-w64-mingw32-gcc.

These wrapper programs strip -mno-cygwin, and convert Windows paths to
Unix paths. They also set LD_PRELOAD and FAKETIME, because those can't
be set for the outer Wine process.

Doing a subprocess.Popen of a Unix program from within Wine mostly
works. For example, subprocess.Popen(["/usr/bin/i686-w64-mingw32-gcc"])
starts a new compiler process. However, if you .wait() on the process
object, you get a "WindowsError: [Error 6] Invalid handle". The same
goes for the implicit wait in subprocess.call and
subprocess.check_output. The .pid of such processes is 0.

What you can do, however, is read from one of the process's output
streams--it will be closed when the process terminates. Because we know
we are dealing with gcc subprocesses, we read its stderr and look for
the string " error: ". If it appears, we return a non-zero status code.
We don't return non-zero merely on a non-empty stderr, because warnings
don't usually cause that.

The difficulty doesn't exist when running native Windows binaries.
That's why we can do subprocess.call(["winepath", ...]): because
winepath.exe is a native binary that comes with Wine.