Compile and install wine-wrappers.
gcc.exe, dllwrap.exe, and swig.exe need to be actual executables. If you
merely rename gcc.py and dllwrap.py, the programs run, but they
immediately detach themselves from the shell and run in the background.
So when py2exe calls a bunch of programs in series:
	gcc -o file1.o file1.c
	dllwrap file1.o
	gcc -o file2.o file2.c
	dllwrap file2.o
	gcc -o file3.o file3.c
	dllwrap file3.o
all the programs would run essentially simultaneously. dllwrap wasn't
waiting for gcc to finish writing the object file that it (dllwrap)
needed as input.

(You could actually get a successful build out of this, if you ran
build_ext once, waiting a few seconds for all the gccs to finish, then
ran it again. The second time, distutils would notice that the object
files already existed, and run only the dllwrap commands, which would
succeed because the object files were already in place.)