Set GIT_COMMITTER_{NAME,EMAIL,DATE} in preparation for applying patches.
These values are included in the commit hash (which is in turn included
in the tor source code as tor_git_revision).

GIT_COMMITTER_NAME is probably always "Ubuntu" in our configuration, and
GIT_COMMITTER_DATE isn't strictly needed as it's taken care of by
libfaketime. But GIT_COMMITTER_EMAIL, if not set, is constructed as
something like "ubuntu@$(hostname --fqdn)", which can vary between hosts
even within the build VM.