Put the "Tor" subdirectory at the head of PATH for DLL access.
Pluggable transport executables are in their own subdirectory,
PluggableTransports, but they need to have access to the OpenSSL DLLs.
Adding the directory containing the DLLs to PATH makes them visible. I
put the directory at the front of PATH and not the end, just in case
there is a libeay32.dll or ssleay32.dll somewhere else on the system.

https://trac.torproject.org/projects/tor/ticket/10845