Install argparse in windows pluggable-transports.
coderman on IRC noticed that the building of obfsproxy in the windows
pluggable transports descriptor was downloading argparse:

	Processing dependencies for obfsproxy==unknown
	Searching for argparse
	Reading https://pypi.python.org/simple/argparse/
	Application tried to create a window, but no driver could be loaded.
	Make sure that your X server is running and that $DISPLAY is set correctly.
	err:systray:initialize_systray Could not create tray window
	fixme:toolhelp:CreateToolhelp32Snapshot Unimplemented: heap list snapshot
	fixme:toolhelp:Heap32ListFirst : stub
	Best match: argparse 1.2.1
	Downloading http://argparse.googlecode.com/files/argparse-1.2.1.tar.gz#md5=2fbef8cb61e506c706957ab6e135840c
	Processing argparse-1.2.1.tar.gz

argparse was already being installed in the linux and mac descriptors.
It was not needed in the windows descriptor, because the windows build
uses Python 2.7, and Python 2.7 has argparse in the standard library;
and the "py2exe" target was able to find the standard module. However
commit 8f8a4cd59bfe07e5b8c12e3990627544d9caeab4 changed obfsproxy to
additionally call the "install" target (so that fteproxy could use
obfsproxy), and apparently "install" can't find argparse. (Maybe because
it's a .py file instead of a directory or an egg?)