Use multiple flashproxy bridge lines.
In https://trac.torproject.org/projects/tor/ticket/10429#comment:2 we
got a bit better reliability with more bridge lines. This doesn't cause
flashproxy-client to send more registration methods, but it does hint to
tor that if a bridge is being slow or unreliable it can try connecting
again to get another.

The corresponding change in the flashproxy repo is
https://gitweb.torproject.org/flashproxy.git/commitdiff/51bdfe59cad37a5f2ffdac3b5f40fe5e2f402a2e.