Bug 10120: Part 1: Refactor Linux descriptors.
We build the tools needed for the Linux descriptors separately and
preserve them. This does not (yet) include the tools in the Pluggable
Transports descriptor which will happen in an other commit.

Additonally, we enhance our clean targets in the Makefile containing not
only a target for the newly created utils component now but for the
other components as well.