Bug 10120: Part 4: Refactor PT descriptors.
We build GMP separately and preserve it now as well. Apart from that
we use the already built OpenSSL in the PT descriptors and leave other
requirements for PTs alone: the overhead building them every time is
not that high and they might be changing more often than the other
utilities we preserve.

Additionally, the descriptors got cleaned-up a bit.