Log Tor Browser output to file and disown the process.
This will log all Tor Browser's output to ``tbb-debug.log`` in the Tor
Browser $HOME directory.

After starting the Firefox process, we grab Firefox's PID and do a dance
where we wait up to 15 seconds to see if Firefox died (which still is
considered to have happened even if the popup window that another Tor
Browser is running has not been exited out of). If Firefox isn't dead,
we disown the process from the shell and report that exit code. If
Firefox did die somewhere, we do:

  wait "$pid"
  exitcode="$?"

in order to get the exitcode from the backgrounded Firefox process. The
call to ``wait`` here doesn't actually wait, because it's only called if
the call to ``kill -0 "$pid"`` exited with non-0 status (meaning that
the process has died).

As a last case fallback, in case something really weird happened, for
example if Firefox didn't die within those 15 seconds that we were
waiting, and then it somehow died all-of-a-sudden right before the

 if test -z "${exitcode}" [...]

line, then we exit with a code 66.

 * CHANGES RelativeLink/RelativeLink.sh to log to tbb-debug.log file,
   and then disown the Firefox process, so that we're not leaving a
   shell open forever (or leaving the shell to decide where the output
   is going to end up).