Bug 12391: Get mingw-w64 compiled reliably.
It seems libfaketime is the culprit for the intermittent build failures
(which might be nothing else than an instance of bug 11459). The
workaround is to omit the compiler (and linker) from this restriction
and to cross the fingers that the libgcc* etc. we ship, too, are still
getting built deterministically. So far this seems to work out...