Set a blackhole proxy in the meek-http-helper profile.
https://trac.torproject.org/projects/tor/ticket/12674

The extension manually overrides the proxy setting for every request, so
this doesn't affect the functioning of the extension. It's intended as a
failsafe to prevent network interaction in case something goes wrong and
the user manages to disable the extension and open the profile with a
normal browser window. In order to cause that to happen, I had to do:

./Browser/firefox -safe-mode -profile Data/Browser/profile.meek-http-helper/

Prior to this commit, running the above command would have given you an
ordinary unproxied Firefox. Now you will get "The proxy server is
refusing connections."