Bug 13359: Revert LD_LIBRARY_PATH + move libstdc++.
Amending the LD_LIBRARY_PATH in our start script breaks the update from
a Tor Browser with the older value to a Tor Browser requiring the new
one. We therefore move the libstdc++ to the directory with the libraries
tor depends upon and LD_LIBRARY_PATH used to point to and change the
LD_LIBRARY_PATH value back to the old one.