Bug #13015: use git tags to set tor-browser version
This commit adds a get-tb-version script which uses the git tags to
output the Tor Browser version and its build directory. The output
should be used with 'eval' in shell scripts to define the variables.
It takes as argument the type of version (release, alpha or beta).

If the current commit is tagged with tag tbb-4.2-build15, get-tb-version
will output this:

    TORBROWSER_VERSION=4.2
    TORBROWSER_BUILDDIR=4.2-build15
    TORBROWSER_SYMLINK_VERSION=1

If the current commit is not tagged, get-tb-version will use the version
from the last tagged commit in history. The build directory will include
the commit hash rather than a build number. The output will be something
like this:

    TORBROWSER_VERSION=4.0
    TORBROWSER_BUILDDIR=4.0-1a98e3efb368
    TORBROWSER_SYMLINK_VERSION=0

The TORBROWSER_BUILDDIR variable is used as the build directory.

The TORBROWSER_SYMLINK_VERSION variable indicates whether we should
create a symlink TORBROWSER_VERSION -> TORBROWSER_BUILDDIR.

The various files which source the versions files and use
TORBROWSER_VERSION have been updated to eval the output of
get-tb-version too, and use TORBROWSER_BUILDDIR as build directory.