Add golang.org/x/crypto to the build process.
This is required for any of:
 * Go >= 1.4.x (It is backwards compatible though).
 * obfs4proxy >= cdeda5724124ca393c87be6d01c84fe4f906d612

And replaces go.crypto (the developers moved/renamed the package).  As
there is no formal release process as far as I can tell for the
codebase the current tip of tree commit is used for versions.alpha, and
master for versions.nightly.

When obfs4proxy 0.0.4 is tagged, the old go.crypto code can be removed
from the build process after the new tag is added to all the
versions.foo files.

Fixes #14316.