Send all output to /dev/null and disown, unless using --debug.
If the start-tor-browser script is called normally, this will send all
output to /dev/null and disown the firefox process.

If --debug is passed to the start-tor-browser script, then output will
go to the terminal it was called from, as well as the logfile,
'tor-browser-debug.log', in the top level of Tor Browser's $HOME
directory.  In this case, we don't disown the child firefox process, and
we also pass the -jsconsole parameter to firefox.

 * FIXES part of #13375: https://bugs.torproject.org/13375