Bug 12468: Send script output messages to stderr only in debug mode
Unity in Ubuntu 14.04 reportedly allocates a pty for use as the stdout and
stderr of all GUI programs (including programs run from Nautilus), and
reportedly does not display messages sent to that pty to the user.

This 'feature' of Ubuntu means that complain can no longer rely on any
terminal to inform the user of an error, unless the user specifically
requests debugging output using a command-line option.

Based on patches from Robert Ransom in
https://trac.torproject.org/projects/tor/ticket/12468.