Bug 15539: Make stripping signatures reproducible
NSIS is neither padding properly nor creating a correct PE-file
checksum. This makes it hard to verify that the Windows installer we
ship is actually matching the one we got built reproducibly.

This fix makes sure that we pad the .exe where needed and that the
PE-file checksum is generated.

Thanks to a cypherpunk for this workaround idea.