Bug 16909: Adapt to HTTPS-Everyhwere build changes
There are two changes that got made recently:
1) The extension ID got renamed due to conflicts on the AMO dashboard
   with the version hosted by Mozilla.
2) Translations moved to an own git submodule.

The first change is unproblematic. For the second one, however, we need
to bump the mininmum Ubuntu version for the host operating system to
14.04. We want to initialize and update the submodules right while we
are fetching the inputs and not during build time to be not dependent
on network access during the bundling step. This fails on Ubuntu 12.04
due to an old git version (1.7.9.5) that uses abolute paths to locate
submodules. As this path is not available on the build VM the build
breaks. This does not affect Debian Wheezy (which ships with
git 1.7.10.4) nor any later Debian version.

In order to make this change obvious for 12.04 users we include an
Ubuntu version check in check-prerequisites.sh.

Additionally, we cleaned the bundling descriptors up to implement a
proper solution for #10066 instead of working around that bug.