Bug 13819: Ship expert bundles with console
Currently, the expert bundles are identical to the tor software we ship
in Tor Browser. This means there is no console showing up displaying
status messages while tor and the browser are running.

This commit fixes this by compiling tor twice, once for Tor Browser
without console showing up and the second time with the console enabled.
The first goes into Tor Browser and the second one is used for the
expert bundle.