Bug 13252: Do not store data in the app bundle
On Mac OS, set DATA_OUTSIDE_APP_DIR=1 for alpha builds. When this
option is enabled, the package layout is altered to be compatible with
Apple's Gatekeeper code signing. All user data will be created in a
TorBrowser-Data directory which sits next to the app bundle and all code
(and only code) is placed under Contents/MacOS.

Currently, FTE is not included when DATA_OUTSIDE_APP_DIR=1 because FTE
needs to be repackaged to separate the Python scripts from the compiled
libraries (see bug 18495).