Bug 18331: Update OS X toolchain for ESR 45
Our old cross-compiler does not cope with ICU changes made in ESR 45. We
switch to the tools Mozilla uses itself instead of creating new ones
based solely on crosstools-ng. This has a number of advantages. The most
important ones are:

1) Using the same toolchain as Mozilla (even if they are not using it
  in official builds) gives us a high degree of confidence that bugfix
  releases will work with our cross-compiler.
2) Using clang both as the host and cross-compiler makes building GCC
  just for being used as host compiler in our Ubuntu Precise environment
  superfluous.

To make it easier compiling clang from source this patch set switches
to Debian Wheezy for the Gitian VMs used to build Tor Browser for OS X.
This fixes #18690.