Upgrade go to 1.6.2.
It now requires an installation of go 1.4.x to bootstrap the build.

We set CGO_ENABLED=0 when building the bootstrap go compiler because go
1.4.3 won't build with a newer GNU toolchain:
https://github.com/golang/go/issues/13114. It didn't cause a problem for
me on Ubuntu 14.04, but this will prevent it from breaking in the
future. We don't need cgo in the bootstrap compiler.

The other change is that the go build system no longer allows us to pass
a command with arguments for CC_FOR_TARGET. I opened a ticket for it:
https://github.com/golang/go/issues/15457. We need -I and -L arguments
in the mac build so that gcc gan find its headers and libraries. So we
wrap up the arguments in a shell script and use the shell script as
CC_FOR_TARGET.