Bug 18331: add /usr/sbin:/sbin to the PATH
On Debian wheezy, dpkg -i fails when the PATH does not include /sbin
and /usr/sbin, so we add them. As a policy, Debian does not include them
by default in regular users PATH.

This change is not required when building using LXC as the commands are
run using sudo which overrides the PATH with the value in "secure_path"
which includes sbin directories.