Bug 18331: drop faketime in mac/gitian-tor.yml
After updating the OSX toolchain, faketime is causing an error in tor's
configure script. As faketime in mac/gitian-tor.yml is no longer
required to get the builds reproducible, we drop it.

Following this change, the intermediary archive tor-mac64-gbuilt.zip
won't be reproducible because of the timestamp of the included files.
However this file is not exposed to users. This should be fixed when we
implement #18845.