Pare down webrtc linux deps.
We avoid some dependencies by not building tests. But one of the tests
is needed because go-webrtc uses the FakeAudioCaptureModule class it
defines. So run ninja twice, the first time with tests disabled to build
the library, and the second time with tests enabled to build just the
libjingle_peerconnection_unittest.isolated target.