Bug 19348: Windows %BUILD_TARGET% changed in Tor Browser 6.0
The values in build_targets can now be a list of values rather than a
single value and are all included in the generated .htaccess file. They
are also all tested in check_update_responses_deployement.

Additionally, a trailing '/' in the last RewriteRule in the .htaccess has
been added. This missing '/' was the reason why the requests using the
new BUILD_TARGET values were returning the en-US complete mar file
rather than an error.