Bug 20184: OS X builds are still not reproducible
We ecountered on some machines failures to reproduce OS X builds.
The investigation seems to indicate that our old toolchain plays a
crucial role in this.

This patch replaces this old toolchain with a more modern clang/cctools
for building the tor part of the bundle. Resorting to `faketime` again
was necessary to cope with inserted timestamps.