Remove webapprt/omni.ja rezipping
After trying to apply the fix for #20147 it turns out that we had an
error in the Firefox descriptor for Windows: the webapprt/omni.ja file
does not exist. There is therefore no need to re-zip it in order to get
a deterministic layout.

The current script we use ignores this error while the new one proposed
in #20147 is not so lenient anymore.